create productAttributeRepository to be able to apply permission on EE
