+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/AttributeRepository.php
@@ -45,27 +45,6 @@ class AttributeRepository extends EntityRepository implements
     /**
      * {@inheritdoc}
      */
-    public function findUniqueAttributeCodes()
-    {
-        $codes = $this
-            ->createQueryBuilder('a')
-            ->select('a.code')
-            ->andWhere('a.unique = ?1')
-            ->setParameter(1, true)
-            ->getQuery()
-            ->getArrayResult();
-
-        return array_map(
-            function ($data) {
-                return $data['code'];
-            },
-            $codes
-        );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function findMediaAttributeCodes()
     {
         $codes = $this
@@ -87,25 +66,6 @@ class AttributeRepository extends EntityRepository implements
             },
             $codes
         );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function findAllAxesQB()
-    {
-        $qb = $this->createQueryBuilder('a');
-        $qb
-            ->andWhere(
-                $qb->expr()->in(
-                    'a.type',
-                    [AttributeTypes::OPTION_SIMPLE_SELECT, AttributeTypes::REFERENCE_DATA_SIMPLE_SELECT]
-                )
-            )
-            ->andWhere($qb->expr()->neq('a.scopable', 1))
-            ->andWhere($qb->expr()->neq('a.localizable', 1));
-
-        return $qb;
     }
 
     /**
@@ -148,75 +108,8 @@ class AttributeRepository extends EntityRepository implements
 
     /**
      * {@inheritdoc}
-     */
-    public function getAttributesAsArray($withLabel = false, $locale = null, array $ids = [])
-    {
-        $qb = $this->_em->createQueryBuilder()
-            ->select('att')
-            ->from($this->_entityName, 'att', 'att.code');
-        if (!empty($ids)) {
-            $qb->andWhere('att.id IN (:ids)')->setParameter('ids', $ids);
-        }
-        $results = $qb->getQuery()->execute([], AbstractQuery::HYDRATE_ARRAY);
-
-        if ($withLabel) {
-            $labelExpr = 'COALESCE(NULLIF(trans.label, \'\'), CONCAT(CONCAT(\'[\', att.code), \']\'))';
-            $groupLabelExpr = 'COALESCE(NULLIF(gtrans.label, \'\'), CONCAT(CONCAT(\'[\', g.code), \']\'))';
-
-            $qb = $this->_em->createQueryBuilder()
-                ->select('att.code', sprintf('%s as label', $labelExpr))
-                ->from($this->_entityName, 'att')
-                ->leftJoin('att.translations', 'trans', 'WITH', 'trans.locale = :locale')
-                ->leftJoin('att.group', 'g')
-                ->leftJoin('g.translations', 'gtrans', 'WITH', 'gtrans.locale = :locale')
-                ->addSelect('g.sortOrder')
-                ->addSelect(sprintf('%s as groupLabel', $groupLabelExpr))
-                ->setParameter('locale', $locale);
-            if (!empty($ids)) {
-                $qb->andWhere('att.id IN (:ids)')->setParameter('ids', $ids);
-            }
-            $attributes = $qb->getQuery()->execute([], AbstractQuery::HYDRATE_ARRAY);
-            foreach ($attributes as $data) {
-                $results[$data['code']]['label'] = $data['label'];
-                $results[$data['code']]['group'] = $data['groupLabel'];
-                $results[$data['code']]['groupOrder'] = $data['sortOrder'];
-            }
-        }
-
-        return $results;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getAttributeIdsUseableInGrid($codes = null, $groupIds = null)
-    {
-        $qb = $this->_em->createQueryBuilder()
-            ->select('att.id')
-            ->from($this->_entityName, 'att', 'att.id');
-
-        if (is_array($codes) && !empty($codes)) {
-            $qb->andWhere("att.code IN (:codes)");
-            $qb->setParameter('codes', $codes);
-        }
-
-        if (is_array($groupIds) && !empty($groupIds)) {
-            $qb->andWhere("att.group IN (:groupIds)");
-            $qb->setParameter('groupIds', $groupIds);
-        } elseif (is_array($groupIds)) {
-            return [];
-        }
-
-        $qb->andWhere('att.useableAsGridFilter = :useableInGrid');
-        $qb->setParameter('useableInGrid', 1);
-
-        $result = $qb->getQuery()->execute([], AbstractQuery::HYDRATE_ARRAY);
-
-        return array_keys($result);
-    }
-
-    /**
-     * {@inheritdoc}
+     *
+     * no need permission
      */
     public function getIdentifier()
     {
@@ -266,6 +159,8 @@ class AttributeRepository extends EntityRepository implements
 
     /**
      * {@inheritdoc}
+     *
+     * used for asset.
      */
     public function getAttributeCodesByType($type)
     {
@@ -286,6 +181,8 @@ class AttributeRepository extends EntityRepository implements
 
     /**
      * {@inheritdoc}
+     *
+     * used to normalize attribute in standard format. no need permission
      */
     public function getAttributeCodesByGroup(AttributeGroupInterface $group)
     {
@@ -306,6 +203,8 @@ class AttributeRepository extends EntityRepository implements
 
     /**
      * {@inheritdoc}
+     *
+     * used to normalize family in standard format. no need permission
      */
     public function findAttributesByFamily(FamilyInterface $family)
     {
@@ -321,6 +220,8 @@ class AttributeRepository extends EntityRepository implements
 
     /**
      * {@inheritdoc}
+     *
+     * used for analytics bundle. no need permission
      */
     public function countAll()
     {
